+++ ChangeLog.md
@@ -13,6 +13,9 @@ Most recent change on the bottom.
 ### Fixed
 - `BatchNorm` was not equivariant for pseudo-scalars
 
+### Added
+- `biases` argument to `o3.Linear`
+
 ## [0.3.2] - 2021-06-10
 ### Added
 - [`opt_einsum_fx`](https://github.com/Linux-cpp-lisp/opt_einsum_fx) as a dependency